+++ test/unit/check_log/log_2_output.py
@@ -45,6 +45,7 @@ class ArgParser(object):
 
     Methods:
         __init__
+        arg_exist
 
     """
 
@@ -59,6 +60,21 @@ class ArgParser(object):
         """
 
         self.args_array = dict()
+
+    def arg_exist(self, arg):
+
+        """Method:  arg_exist
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_default.
+
+        Arguments:
+
+        """
+
+        if arg in self.args_array:
+            return True
+
+        return False 
 
 
 class UnitTest(unittest.TestCase):