+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [4.0.0] - 2022-02-07
+Breaking Changes
+
+- Set up instance of the gen_class.ArgParser class which replaces all manual operations of the args_array variable that contain the command line arguments.
+
+
 ## [3.1.4] - 2021-11-30
 -  Allow to override the default sendmail (postfix) and use mailx command.